requirements.txt
asgiref==3.5.2
boto3==1.25.5
botocore==1.28.5
Django==4.1.2
djangorestframework==3.14.0
djangorestframework-gis==1.0
django-allow-cidr==0.5.0
django-crispy-forms==1.14.0
django-mathfilters==1.0.0
django-storages==1.13.1
django-utils-six==2.0
django-cleanup==8.0.0
gunicorn==20.1.0
jmespath==1.0.1
packaging==21.3
Pillow==9.2.0
psycopg2-binary==2.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.5
urllib3==1.26.12
geopy==2.3.0
